Include flat directory in fuse paths
Prior to this, the 'girder mount' command exposes some girder resources
as
<mount_path>/<user_or_collection>/<folder>/[<folder>/...]<item>/<file>.
This allows libraries that require actual files to read remote assets as
if they are local files.  However, since girder includes items in the
hierarchy, libraries that read files relative to one another often
cannot find the adjacent files.

This PR exposes an alternative set of paths without removing the
original behavior.  Specifically, files also are exposed as
<mount_path>/flat/<user_or_collection>/<folder>/[<folder>/...]/<file>.
Since items can have multiple files, this only exposes one file per
item.  If the item has multiple files, it picks the earliest id.  This
could be overridden by a plugin if there was some reason to have a
different ordering.  Similarly, if there are no files in an item, the
item remains represented as an empty directory.

There is a convenience method to get the local mount path; this has an
option to prefer the flat path.  It does a nominal check that the file
requested is the same file as is exposed in the flat path, but this is a
cheap check based on file size since the information is not otherwise
available and a true check would be expensive.

The motivation for this change is to allow reading multi-file image
formats via large_image that are imported from S3 and use an opaque
library like bioformats for reading the image.